feat: add test and refacto
